+++ achievements/achievements.go
@@ -7,9 +7,6 @@ import (
 
 	"github.com/framey-io/go-tarantool"
 	"github.com/hashicorp/go-multierror"
-	achievementprocessor "github.com/ice-blockchain/santa/achievements/internal/achievement-processor"
-	economy_processor "github.com/ice-blockchain/santa/achievements/internal/economy-processor"
-	user_processor "github.com/ice-blockchain/santa/achievements/internal/user-processor"
 	appCfg "github.com/ice-blockchain/wintr/config"
 	messagebroker "github.com/ice-blockchain/wintr/connectors/message_broker"
 	"github.com/ice-blockchain/wintr/connectors/storage"
@@ -37,8 +34,7 @@ func StartProcessor(ctx context.Context, cancel context.CancelFunc) Processor {
 	appCfg.MustLoadFromKey(applicationYamlKey, &cfg)
 	db := storage.MustConnect(ctx, cancel, ddl, applicationYamlKey)
 	mbProducer := messagebroker.MustConnect(context.Background(), applicationYamlKey)
-	repo := &repository{db: db, mb: mbProducer}
-	mbConsumer := messagebroker.MustConnectAndStartConsuming(context.Background(), cancel, applicationYamlKey, processors(repo, db))
+	mbConsumer := messagebroker.MustConnectAndStartConsuming(context.Background(), cancel, applicationYamlKey, processors(mbProducer, db))
 
 	return &processor{
 		close: func() error {
@@ -61,18 +57,16 @@ func StartProcessor(ctx context.Context, cancel context.CancelFunc) Processor {
 				return multierror.Append(nil, result...)
 			}
 		},
-		WriteRepository: repo,
 	}
 }
 
-func processors(repo WriteRepository, db tarantool.Connector) map[messagebroker.Topic]messagebroker.Processor {
+func processors(mb messagebroker.Client, db tarantool.Connector) map[messagebroker.Topic]messagebroker.Processor {
 	return map[messagebroker.Topic]messagebroker.Processor{
-		// May be it is better to iternate and look for topics name?
-		// Because of current impementation requires topic to be in specific order in configuration.
-		cfg.MessageBroker.ConsumingTopics[0]: user_processor.New(db, repo),
-		cfg.MessageBroker.ConsumingTopics[1]: economy_processor.NewMiningEventProcessor(db, repo),
-		cfg.MessageBroker.ConsumingTopics[2]: achievementprocessor.NewTaskProcessor(db, repo),
-		cfg.MessageBroker.ConsumingTopics[3]: achievementprocessor.NewBadgeProcessor(db),
+		// TODO: fill with new processors
+		//cfg.MessageBroker.ConsumingTopics[0]: user_processor.New(db, repo),
+		//cfg.MessageBroker.ConsumingTopics[1]: economy_processor.NewMiningEventProcessor(db, repo),
+		//cfg.MessageBroker.ConsumingTopics[2]: achievementprocessor.NewTaskProcessor(db, repo),
+		//cfg.MessageBroker.ConsumingTopics[3]: achievementprocessor.NewBadgeProcessor(db),
 	}
 }
